pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG system for Nerd-o"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "agno",
    "python-dotenv",
    "openai",
    "sqlalchemy",
    "psycopg",
    "psycopg2-binary",
    "pgvector",
    "streamlit",
    "pypdf",
    "chonkie",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
